apptrack-backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
Pillow==10.2.0
gunicorn==21.2.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
pre-commit==3.6.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
whitenoise==6.6.0
factory-boy==3.3.0
freezegun==1.4.0